+++ Configuration GUI/src/common/Validators.ts
@@ -18,7 +18,7 @@
  * Fabian Ihle (fabian.ihle@uni-tuebingen.de)
  */
 
-import { ASIC, DefaultStream, DefaultStreamSettings, MPLSHeader, PortInfo, Stream, StreamSettings } from "./Interfaces";
+import { ASIC, DefaultStream, DefaultStreamSettings, MPLSHeader, P4TGInfos, PortInfo, Stream, StreamSettings } from "./Interfaces";
 
 export const validateMAC = (mac: string) => {
     let regex = /^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$/;
@@ -123,14 +123,16 @@ export const validateStreams = (s: Stream[]) => {
     return s.every(s => Object.keys(defaultStream).every(key => Object.keys(s).includes(key)))
 }
 
-export const validatePorts = (port_tx_rx_mapping: { [name: number]: number }, available_ports: PortInfo[]) => {
+export const validatePorts = (port_tx_rx_mapping: { [name: number]: number }, available_ports: PortInfo[], p4tg_infos: P4TGInfos) => {
     // Verify if all configured ports are acutally available on this device.
+    const available_ports_for_tg = available_ports
+        .filter(port => port["loopback"] === "BF_LPBK_NONE" || p4tg_infos.loopback)
+        .map(port => port.port);
 
-    const available_dev_ports: number[] = available_ports.slice(0, 10).map(p => p.pid);
     //@ts-ignore
     const configured_ports: number[] = Object.entries(port_tx_rx_mapping).flatMap(([key, value]) => [Number(key), value]);
 
-    return configured_ports.length == 0 || configured_ports.every(r => available_dev_ports.includes(r))
+    return configured_ports.length == 0 || configured_ports.every(r => available_ports_for_tg.includes(r))
 }
 
 export const validateStreamSettings = (setting: StreamSettings[]) => {